move MetricLogPathBuilder into export/type.go
