+++ pkg/util/export/type.go
@@ -16,11 +16,16 @@ package export
 
 import (
 	"context"
+	"fmt"
 	"io"
+	"path"
 	"reflect"
+	"strings"
 	"sync/atomic"
+	"time"
 	"unsafe"
 
+	"github.com/matrixorigin/matrixone/pkg/common/moerr"
 	"github.com/matrixorigin/matrixone/pkg/util/batchpipe"
 )
 
@@ -84,4 +89,146 @@ func String2Bytes(s string) (ret []byte) {
 	sliceHead.Len = strHead.Len
 	sliceHead.Cap = strHead.Len
 	return
+}
+
+type MergeLogType string
+
+const MergeLogTypeMerged MergeLogType = "merged"
+const MergeLogTypeLog MergeLogType = "log"
+
+const FilenameSeparator = "_"
+const CsvExtension = ".csv"
+
+// PathBuilder hold strategy to build filepath
+type PathBuilder interface {
+	Clone() PathBuilder
+	// Build directory path
+	Build(string /*account*/, MergeLogType, time.Time, batchpipe.HasName) string
+	// DirectoryPath return last call Build result
+	DirectoryPath() string
+	// Join return DirectoryPath + "/" + filename
+	Join(filename string) string
+	// ParsePath
+	//
+	// switch path {
+	// case "{timestamp_writedown}_{node_uuid}_{ndoe_type}.csv":
+	// case "{timestamp_start}_{timestamp_end}_merged.csv"
+	// }
+	ParsePath(path string) (CSVPath, error)
+	NewMergeFilename(timestampStart, timestampEnd string) string
+}
+
+type CSVPath interface {
+	Table() string
+	Timestamp() []string
+}
+
+var _ CSVPath = (*MetricLogPath)(nil)
+
+type MetricLogPath struct {
+	// path raw data
+	path string
+	// table parsed from path
+	table string
+	// filename
+	filename string
+	// timestamps parsed from filename in path
+	timestamps []string
+	// fileType, val in [log, merged]
+	fileType MergeLogType
+}
+
+const PathElems = 7
+const PathIdxFilename = 6
+const PathIdxTable = 5
+const PathIdxAccount = 0
+const FilenameElems = 3
+const FilenameIdxType = 2
+
+// NewMetricLogPath
+//
+// path like: sys/[log|merged]/yyyy/mm/dd/table/***.csv
+// ##    idx: 0   1            2    3  4  5     6
+// filename like: {timestamp}_{node_uuid}_{node_type}.csv
+// ##         or: {timestamp_start}_{timestamp_end}_merged.csv
+func NewMetricLogPath(path string) *MetricLogPath {
+	return &MetricLogPath{path: path}
+}
+
+func (p *MetricLogPath) Parse() error {
+	// parse path => filename, table
+	elems := strings.Split(p.path, "/")
+	if len(elems) != PathElems {
+		return moerr.NewInternalError("metric/log invalid path: %s", p.path)
+	}
+	p.filename = elems[PathIdxFilename]
+	p.table = elems[PathIdxTable]
+
+	// parse filename => fileType, timestamps
+	filename := strings.Trim(p.filename, CsvExtension)
+	fnElems := strings.Split(filename, FilenameSeparator)
+	if len(fnElems) != FilenameElems {
+		return moerr.NewInternalError("metric/log invalid filename: %s", p.path)
+	}
+	if fnElems[FilenameIdxType] == string(MergeLogTypeMerged) {
+		p.fileType = MergeLogTypeMerged
+		p.timestamps = fnElems[:2]
+	} else {
+		p.fileType = MergeLogTypeLog
+		p.timestamps = fnElems[:1]
+	}
+
+	return nil
+}
+
+func (p *MetricLogPath) Table() string {
+	return p.table
+}
+
+func (p *MetricLogPath) Timestamp() []string {
+	return p.timestamps
+}
+
+var _ PathBuilder = (*MetricLogPathBuilder)(nil)
+
+type MetricLogPathBuilder struct {
+	directory string
+}
+
+func (m *MetricLogPathBuilder) Clone() PathBuilder {
+	builder := NewMetricLogPathBuilder()
+	builder.directory = m.directory
+	return builder
+}
+
+func NewMetricLogPathBuilder() *MetricLogPathBuilder {
+	return &MetricLogPathBuilder{}
+}
+
+func (m *MetricLogPathBuilder) Build(account string, datatype MergeLogType, timestamp time.Time, table batchpipe.HasName) string {
+	m.directory = path.Join(account,
+		string(datatype),
+		fmt.Sprintf("%d", timestamp.Year()),
+		fmt.Sprintf("%02d", timestamp.Month()),
+		fmt.Sprintf("%02d", timestamp.Day()),
+		table.GetName(),
+	)
+	return m.directory
+}
+
+func (m MetricLogPathBuilder) DirectoryPath() string {
+	return m.directory
+}
+
+func (m *MetricLogPathBuilder) Join(filename string) string {
+	return path.Join(m.directory, filename)
+}
+
+func (m *MetricLogPathBuilder) ParsePath(path string) (CSVPath, error) {
+	p := NewMetricLogPath(path)
+	return p, p.Parse()
+}
+
+func (m *MetricLogPathBuilder) NewMergeFilename(timestampStart, timestampEnd string) string {
+	return strings.Join([]string{timestampStart, timestampEnd, string(MergeLogTypeMerged)}, FilenameSeparator) + CsvExtension
 }
